modified:   Makefile 	modified:   ft_strlcpy.c 	modified:   ft_strlen.c 	modified:   ft_strtrim.c 	new file:   ft_substr.c 	modified:   libft.a 	modified:   libft.h 	From 42 Le Havre

+++ ft_strlen.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_strlen(const char *string)
+size_t	ft_strlen(const char *string)
 {
 	int	i;
